remove useless functions

+++ src/ft_pow.c
@@ -6,13 +6,14 @@
 /*   By: qle-guen <qle-guen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/26 16:24:29 by qle-guen          #+#    #+#             */
-/*   Updated: 2016/01/06 11:53:34 by qle-guen         ###   ########.fr       */
+/*   Updated: 2016/06/28 14:53:44 by qle-guen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <libft.h>
 
-int			ft_pow(int a, int n)
+unsigned long	ft_pow
+	(unsigned long a, unsigned long n)
 {
 	if (n == 0)
 		return (1);
